requirements.txt
openai
python-dotenv
streamlit
pybase64